Ganti nama file + update
Penambahan fitur warna, dan nilai random yang dihasilkan selalu berbeda dari nilai sebelumnya
